Fix class / subclass attributes forking
Prior to this commit, attributes were not inherited by a subclass if
they were added to the parent class after the subclass had added its own
attributes.  This was due to the way attribute definitions were stored
in an append-only `class_attribute`.  With this commit, attribute
definitions are stored and iterated in a way that ensures they will
always be inherited.