Fix class / subclass attribute decorations forking
Prior to this commit, attribute decorations were not inherited by a
subclass if they were added to the parent class after the subclass had
added its own attribute decorations.  This was due to the way attribute
decorations were stored in an append-only `class_attribute`.  With this
commit, attribute decorations are stored and iterated in a way that
ensures they will always be inherited.