Run `rails new` using specified prerelease version
When running `rails new` with a prerelease flag like `--edge`, the
project's Gemfile will point to the specified prerelease version, but
the project itself is generated using the active version of Rails.  This
can cause problems if, for example, the prerelease version expects a
different file structure than the active version.  It also precludes
using generator options that are new in the prerelease version.

This commit adds a mechanism to install and use prerelease versions of
Rails when running `rails new`.

Closes #38761.

Co-authored-by: Nick Schwaderer <nicholas.schwaderer@gmail.com>