Listen on fork in EventedFileUpdateChecker
The Listen gem does not notify across process forks, and so forked
processes must re-register their listeners.  `EventedFileUpdateChecker`
handled this by re-registering in `updated?` if `Process.pid` had
changed.  Because file updates might have occurred between the start of
the fork and the call to `updated?`, `updated?` was set to always return
true in this case.

However, the above approach can result in unnecessary app reloading,
particularly when using Spring.  That, in turn, can cause unexpected and
problematic behavior: #39431, #37591.  This commit changes
re-registration to occur immediately on fork, and changes `updated?` to
return true only when necessary.