Load spring/binstub directly in boot.rb
In #39632, `boot.rb` was changed to load `bin/spring`, with the
intention of adding a guard to Spring itself that would prevent Spring
from running in production environments.  However, in a production
environment, the Spring gem may not be installed.  Furthermore, the
contents of `bin/spring` may not be guaranteed as it can be overwritten
by e.g. `bundle binstubs`, and so it can raise unexpected errors.

Therefore, this commit modifies `boot.rb` to include the (code-golfed)
contents of `bin/spring` directly.  Additionally, to keep these two
files in sync, `bin/spring` is now generated by Rails itself rather than
via the `spring binstub` command.

Incidentally, generating `bin/spring` via Rails itself also addresses an
issue with the `--skip-bundle` flag.  Previously, `bin/spring` was not
generated when using `--skip-bundle`, and the user would have to run
both `bundle install` *and* `bundle exec spring binstub` manually,
though the latter step was not documented or explained.