Fix `thread_mattr_accessor` `default` option behavior
This makes the value supplier to the `default` option of
`thread_mattr_accessor` to be set in descendant classes
as well as in any new Thread that starts.

Previously, the `default` value provided was set only at the
moment of defining the attribute writer, which would cause
the attribute to be uninitialized in descendants and in other threads.

For instance:

    class Processor
      thread_mattr_accessor :mode, default: :smart
    end

    class SubProcessor < Processor
    end

    SubProcessor.mode # => :smart
    Thread.new do
      Processor.mode # => :smart
    end.join

when in the past, those two calls would have returned `nil`.

This logic comes with a performance impact on the reader,
here is the benchmark:

    Warming up --------------------------------------
            original   238.780k i/100ms
     patched_default   162.765k i/100ms
    Calculating -------------------------------------
            original    2.376M (± 9.4%) i/s -  11.939M in 5.078355s
     patched_default    1.635M (±16.4%) i/s -   7.813M in 5.008433s

    Comparison:
            original:  2375953.1 i/s
     patched_default:  1634668.2 i/s - 1.45x  (± 0.00) slower

Fixes #43312.