Document database query caveat for authenticate_by [ci-skip]
We are not in a position to make guarantees about how long
`authenticate_by`'s database query will take.  For example, if the
queried attributes are not indexed, the database will need to perform a
full table scan, and query time will vary based on where the record is
in the heap, if it is found at all.  Therefore, we should document this
caveat.