Factor out ActiveModel::AttributeRegistration
As a step toward sharing more code between Active Model and Active
Record, this commit factors an `ActiveModel::AttributeRegistration`
module out of `ActiveModel::Attributes`.  This module is marked as
`nodoc` and is for internal use only.

Additionally, this commit adds thorough test coverage of attribute
registration and inheritance.  (`activemodel/test/cases/attributes_test.rb`
does already test some of this behavior, but it is focused on high-level
functionality.)