Add redirect_code_for_unsafe_http_methods config
When a client follows an HTTP 302 redirect, it will typically use the
same HTTP method as the original request.  This can cause issues when,
for example, redirecting an XHR `DELETE` request after a `destroy`.
(Note that many clients make an exception for `POST`, and will use `GET`
to follow a 302 redirect in that case.)

The solution so far has been for users to specify an explicit response
code in such cases.  For example:

```ruby
def destroy
  Post.destroy(params[:id])
  redirect_to root_path, status: 303
end
```

This commit adds a `redirect_code_for_unsafe_http_methods` configuration
setting that allows users to specify a default HTTP response code to use
when redirecting a request made with an [unsafe HTTP method][], such as
`POST` or `DELETE`.  For example, when set to 303, the explicit response
code may be omitted:

```ruby
def destroy
  Post.destroy(params[:id])
  redirect_to root_path
end
```

[unsafe HTTP method]: https://developer.mozilla.org/en-US/docs/Glossary/Safe/HTTP