Unify LTR / RTL CSS for guides
When `main.css` is updated, it is easy for `main.rtl.css` to be
overlooked (see #45423).

This commit eliminates `main.rtl.css` in favor of a unified approach to
LTR / RTL styling in `main.css`.  To accomplish this, the `<body>`
element is rendered with a [`dir` attribute][] (which has the same
effect as setting the `direction` CSS property), and LTR- / RTL-specific
styles are prefixed with `body[dir="..."]` selectors.

This also paves the way for automatic detection and application of
LTR / RTL styles when the [`:dir()` CSS pseudo-class][] gains widespread
browser support.

[`dir` attribute]: https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/dir
[`:dir()` CSS pseudo-class]: https://developer.mozilla.org/en-US/docs/Web/CSS/:dir