Bucket associations when preloading
This commit modifies the `Preloader` to track preloaded associations in
buckets, and, when preloading a downstream association, to preload that
association for all owner records from the same bucket.

The intent of this change is to allow `includes` and `preload` calls to
be placed alongside the code that depends on them.  This makes it easier
to add a preload when needed or remove a preload when it is no longer
needed.

Consider the following code:

  ```ruby
  def render_authors(authors)
    authors.includes(:posts).each do |author|
      render_posts(author.posts)
    end
  end

  def render_posts(posts)
    posts.includes(:comments).each do |post|
      post.comments.each do |comment|
        # ...
      end
    end
  end

  render_authors(Author.all)
  ```

Before this commit, with N authors each having at least 1 post, the code
would execute N + 2 queries.  For example, with 5 authors:

  ```
  SELECT "authors".* FROM "authors"
  SELECT "posts".* FROM "posts" WHERE "posts"."author_id" IN (?, ?, ?, ?, ?)  [["author_id", 1], ["author_id", 2], ["author_id", 3], ["author_id", 4], ["author_id", 5]]
  SELECT "comments".* FROM "comments" WHERE "comments"."post_id" = ?  [["post_id", 11]]
  SELECT "comments".* FROM "comments" WHERE "comments"."post_id" = ?  [["post_id", 12]]
  SELECT "comments".* FROM "comments" WHERE "comments"."post_id" = ?  [["post_id", 13]]
  SELECT "comments".* FROM "comments" WHERE "comments"."post_id" = ?  [["post_id", 14]]
  SELECT "comments".* FROM "comments" WHERE "comments"."post_id" = ?  [["post_id", 15]]
  ```

After this commit, the code executes only 3 queries:

  ```
  SELECT "authors".* FROM "authors"
  SELECT "posts".* FROM "posts" WHERE "posts"."author_id" IN (?, ?, ?, ?, ?)  [["author_id", 1], ["author_id", 2], ["author_id", 3], ["author_id", 4], ["author_id", 5]]
  SELECT "comments".* FROM "comments" WHERE "comments"."post_id" IN (?, ?, ?, ?, ?)  [["post_id", 11], ["post_id", 12], ["post_id", 13], ["post_id", 14], ["post_id", 15]]
  ```