Dup non-frozen thread_mattr_accessor defaults
This provides a basic level of protection against different threads
trying to mutate a shared default value.  It is not a bulletproof
solution however, because the default value may contain nested mutable
objects.

---

Perhaps it would be better to use `deep_dup` here?  Though that would still not be bulletproof.  For example, when the default is frozen but contains non-frozen objects, the `deep_dup` would be skipped.

Another (not bulletproof) option would be to `.dup.freeze` any non-frozen default when it's declared.  Then threads would be expected to `dup` the value before modifying or use something like `+=` instead.

Or, another option would be to only allow frozen defaults, i.e. raise `ArgumentError` if a non-frozen default is specified.  That would let the user know what is expected, and would (arguably) put the responsibility for corner cases, such as nested mutable objects, on them.