Move SerializeCastValue to Type::Helpers module
This commit moves `ActiveModel::Type::SerializeCastValue` to
`ActiveModel::Type::Helpers::SerializeCastValue`.

Generally, this appears to be a good pattern for extending the contract
of `Type::Value`: (1) implement the extension as a `Helper` sub-module
that is included in `Type::Value`, and (2) implement a `Helper` module
singleton method that handles the case where the given type does not
implement the extension (e.g. because it is not a subclass of
`Type::Value`).