Use memoized dup of url_helpers for reinclusion
In the case where a controller subclasses an engine's controller that,
in turn, subclasses a controller that includes the application's
`url_helpers` (for example, in the "isolated engine routes and helpers
are isolated to that engine" test in `railties/test/railties/engine_test.rb`),
this commit avoids allocating a new module per controller:

  ```ruby
  ActionController::Base.include Rails.application.routes.url_helpers
  C1 = Class.new(ActiveStorage::DirectUploadsController)
  C2 = Class.new(ActiveStorage::DirectUploadsController)

  C1.ancestors - C2.ancestors
  # BEFORE:
  # => [C1, #<Module:0x...>]
  # AFTER:
  # => [C1]
  ```

This commit also reworks the `RouteHelperIntegrationTest` test to use
the controllers defined in `actionpack/test/abstract_unit.rb`.
Otherwise, `extend AbstractController::Railties::RoutesHelpers.with(...)`
happens twice -- once for `ActionController::Base` and once for
`FakeACBase` -- which causes descendents of `FakeACBase` to include an
extra module as they flip-flop their `_routes` definitions.  Before this
commit, the extra module only defined a `_routes` method; with the
change from this commit, the extra module would be the memoized dup of
`routes.url_helpers`, which would cause the "only includes one module
with route helpers" test to falsely fail.