Add Message{Encryptors,Verifiers}#transitional
This commit adds a `transitional` attribute to
`ActiveSupport::MessageEncryptors` and `ActiveSupport::MessageVerifiers`.
Setting `transitional = true` will swap the first two rotations when
building a message encryptor / verifier.  For example, with the
following configuration, message verifiers will generate messages using
`serializer: Marshal, url_safe: true`, and will able to verify messages
that were generated using any of the three option sets:

  ```ruby
  verifiers = ActiveSupport::MessageVerifiers.new { ... }
  verifiers.rotate(serializer: JSON, url_safe: true)
  verifiers.rotate(serializer: Marshal, url_safe: true)
  verifiers.rotate(serializer: Marshal, url_safe: false)
  verifiers.transitional = true
  ```

This can be useful when performing a rolling deploy of an application,
wherein servers that have not yet been updated must still be able to
verify messages from updated servers.  In particular, as it can be
applied to the default rotations of `Rails.application.message_verifiers`.