Support replacing cache compressor
This commit adds support for replacing the compressor used for
serialized cache entries.  Custom compressors must respond to `deflate`
and `inflate`.  For example:

  ```ruby
  module MyCompressor
    def self.deflate(string)
      # compression logic...
    end

    def self.inflate(compressed)
      # decompression logic...
    end
  end

  config.cache_store = :redis_cache_store, { compressor: MyCompressor }
  ```

Additionally, as part of this work, a new optimization has been added
for cache format version 7.1 such that the version and expiry of cache
entries can be checked without deserializing (or decompressing) the
entries.

Also as part of this work, custom cache coders now support compression
when the cache format version is >= 7.1, and they benefit from the
improvements introduced by cache format version 7.1 (namely, lazy
deserialization from this commit, and optimized serialization of bare
strings from #48122).